# backend/docs/API_Documentation.yaml
openapi: 3.0.3
info:
  title: FiBo API Documentation
  description: This is the Documentation for the API that is provided by our Django Backend
  version: 1.0.0
servers:
  - url: 'http://127.0.0.1:8000'
...
paths:
...
  /users/get/:
    get:
      tags:
        - users
      summary: Gets user data from database
      description: '`Returns` the data from the database corresponding to the user that is sending the request.'
      operationId: getUser
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/auth'
      responses:
        '200':
          $ref: '#/components/responses/getUser'
        '401':
          $ref: '#/components/responses/accessError'
  /users/delete/:
    delete:
      tags:
        - users
      summary: Deletes an existing user
      description: '`Deletes` the user from the Database, that is sending the request.'
      operationId: deleteUser
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/auth'
      responses:
        '200':
          description: The Account corresponding to the send Access Token is deleted
        '401':
          $ref: '#/components/responses/accessError'
...
components:
  schemas:
    users:
      required:
        - id
        - email
        - password
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          format: email
          example: theUser@fibo.de
        email:
          type: string
          format: email
          example: theUser@fibo.de
        password:
          type: string
          example: 'password'
        last_login:
          type: string
          format: date
          example: "2023-04-20 07:56:12.809812+00"
        is_superuser:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
        date_joined:
          type: string
          format: date
          example: "2023-04-15 12:23:59.602445+00"
...
  responses:
...
    getUser:
      description: Successful operation. The User corresponding to the send Access Token is returned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/users'
    accessError:
      description: Access token is missing or invalid
...
  parameters:
    auth:
      name: 'Authorization'
      in: header
      description: 'The `Access Token` to be passed as a header'
      required: true
      schema:
        type: string
        example: "Authorization: Bearer <token>"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT